requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
filemagic==1.6
    # via helix (setup.py)
